Hits to data.tables

A generic function for converting Elasticsearch docs into R data.tables. It uses fromJSON with flatten = TRUE to convert a JSON into an R data.frame, and formats it into a data.table.

chomp_hits(hits_json = NULL, keep_nested_data_cols = FALSE)

Arguments

hits_json
A character vector. If its length is greater than 1, its elements will be pasted together. This can contain a JSON returned from a search query in Elasticsearch, or a filepath or URL pointing at one. Most commonly, this JSON will be the direct result of running Search with raw = TRUE.

keep_nested_data_cols
a boolean (default FALSE); whether to keep columns that are nested arrays in the original JSON. A warning will be given if these columns are deleted.

Examples

# NOT RUN {
# Get 100 purchase records

# Get result from Elasticsearch
result <- uptasticsearch::es_search(es_host = "http://es.custdb.mycompany.com:9200"
                                    , es_index = "ticket_sales"
                                    , trailing_args = "size=100")

# Parse into a data.table
resultDT <- chomp_hits(hits_json = result)
# }