Unpack a nested data.table

After calling a chomp_* function or es_search, if you had a nested array in the JSON, its corresponding column in the resulting data.table is a data.frame itself (or a list of vectors). This function expands that nested column out, adding its data to the original data.table, and duplicating metadata down the rows as necessary.

unpack_nested_data(chomped_df, col_to_unpack)

Arguments

chomped_df
a data.table

col_to_unpack
a character vector of length one: the column name to unpack

Examples

# Chomp a dummy sample JSON of hits
sampleChompedDT <- chomp_hits('[{"_source":{"timestamp":"2017-01-01",
  "cust_name":"Austin", "details":{"cust_class":"big_spender","location":
  "chicago","pastPurchases":[{"film":"The Notebook", "pmt_amount":6.25},
  {"film":"The Town", "pmt_amount":8.00},{"film":"Zootopia","pmt_amount": 
  7.50, "matinee": true}]}}},{"_source": {"timestamp":"2017-02-02",
  "cust_name":"James","details":{"cust_class":"peasant","location":"chicago",
  "pastPurchases":[{"film":"Minions", "pmt_amount":6.25, "matinee": true},
  {"film":"Rogue One","pmt_amount":10.25},{"film":"Bridesmaids",
  "pmt_amount":8.75},{"film":"Bridesmaids","pmt_amount":6.25, "matinee": 
  true}]}}}]'
  , keep_nested_data_cols = TRUE)

#> INFO [2017-07-14 15:03:41] Keeping the following nested data columns. Consider using unpack_nested_data for one:
#>  details.pastPurchases

# Unpack by details.pastPurchases
unpackedDT <- unpack_nested_data(chomped_df = sampleChompedDT
                                 , col_to_unpack = "details.pastPurchases")